19/02/2024

+++ src/main/java/com/chat/websocket/repository/GroupMemberRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.Optional;
 
-public interface GroupMemberRepository extends JpaRepository<GroupMember,Integer> {
+public interface GroupMemberRepository extends JpaRepository<GroupMember,Long> {
     @Query("select gm from GroupMember gm where gm.contact.email = :email and gm.conversation.id = :conversationId" )
     Optional<GroupMember> findByEmailAndConversationId(@Param("email") String email,@Param("conversationId") int conversationId);
 }
